- Little Commit

+++ Testing/Engine/Shaders/GameView.vert
@@ -1,5 +1,6 @@
-#version 330 core
-layout (location = 0) in vec3 aPos;
+#version 450 core
+
+layout (location = 0) in dvec3 aPos;
 layout (location = 1) in vec3 aNormal;
 
 out vec3 Normal;
